+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - (Feature) (Platform) Platform Requirements support
 - (Improvement) Drop slash requirement from ArangoRoute
 - (Feature) (Networking) Pass through Server Header
+- (Feature) (Platform) Shutdown migration to CE
 
 ## [1.2.43](https://github.com/arangodb/kube-arangodb/tree/1.2.43) (2024-10-14)
 - (Feature) ArangoRoute CRD